CLIArgumentParser {pkgmaker}	R Documentation

Enhanced Command Line Argument Parser

Description

Extends the capabilities of package argparse, e.g., in defining sub commands.

parseCMD parse command line arguments for sub-commands, and dispatch to the associated function.

Usage

  CLIArgumentParser(prog = CLIfile(), description = "",
    ..., epilog = "", show.defaults = TRUE)

  parseCMD(parser, ARGS = commandArgs(TRUE), debug = FALSE,
    envir = parent.frame())

Arguments

prog
program name

description
program description

...
extra arguments passed to ArgumentParser.

epilog
epilog messages to display at the end of the man pages

show.defaults
logical that indicates if default arugment values should be displayed.

parser
parser object as returned by CLIArgumentParser.

ARGS
command line argument to parse, as a named list or a character string.

debug
logical that indicate if debugging information should be printed.

envir
environment that contains where the sub-command functions are looked for.

[Package pkgmaker version 0.22 Index]