R.exec {pkgmaker}	R Documentation

Executing R Commands

Description

R.exec executes a single R command via system2.

R.CMD executes R CMD commands.

R.SHLIB executes R CMD SHLIB commands.

Usage

  R.exec(..., lib.loc = NULL)

  R.CMD(cmd, ...)

  R.SHLIB(libname, ...)

Arguments

...
extra arguments that are concatenated and appended to the command.

lib.loc
logical that indicates if the current library locations should be used. If a character vector, then it is used as the library path specification.

cmd
command to run, e.g. ‘check’ or ‘INSTALL’.

libname
name of the output compiled library

[Package pkgmaker version 0.22 Index]