RdSection2latex {pkgmaker}	R Documentation

Format Rd Sections into LatTeX

Description

This function extract sections from Rd files and convert them into LaTeX code. This can be useful to include Rd text into vignettes, hence keeping them up to date.

Usage

  RdSection2latex(topic, package, i = 1L, notitle = TRUE)

Arguments

topic
Rd topic

package
package in which to search the topic

i
index of the section to format

notitle
logical that indicates if the section's title should be removed

Example section

This is a nice section, with a bullet list:

tata

toto

Examples

RdSection2latex('RdSection2latex', package = 'pkgmaker')

[Package pkgmaker version 0.22 Index]