Sys.getenv_value {pkgmaker}	R Documentation

System Environment Variables

Description

System Environment Variables

Usage

  Sys.getenv_value(name, raw = FALSE)

Arguments

name
variable name as a character string.

raw
logical that indicates if one should return the raw value or the convertion of any false value to FALSE.

Value

the value of the environment variable as a character string or NA is the variable is not defined at all.

Examples

# undefined returns FALSE
Sys.getenv_value('TOTO')
# raw undefined returns NA
Sys.getenv_value('TOTO', raw = TRUE)

Sys.setenv(TOTO='bla')
Sys.getenv_value('TOTO')

# anything false-like returns FALSE
Sys.setenv(TOTO='false'); Sys.getenv_value('TOTO')
Sys.setenv(TOTO='0'); Sys.getenv_value('TOTO')

# cleanup
Sys.unsetenv('TOTO')

[Package pkgmaker version 0.22 Index]