addToLogger {pkgmaker}	R Documentation

Enhancing RUnit Logger

Description

Adds a function or a local variable to RUnit global logger.

Usage

  addToLogger(name, value, logger = NULL)

Arguments

name
name of the function or variable to add

value
object to append to the logger. If value is a function it is added to the list and is accessible via .testLogger$name. If value is a variable it is added to the local environment and is therefore accessible in all logging functions.

logger
an optional RUnit logger object. If missing or NULL, the object .testLogger is searched in .GlobalEnv – and an error is thrown if it does not exist.

Value

the modified logger object. Note that the global object is also modified if logger is NULL.

[Package pkgmaker version 0.22 Index]