addnames {pkgmaker}	R Documentation

Generating Names

Description

Generates names or dimnames for objects.

Usage

  addnames(x, ...)

  ## Default S3 method:
 addnames(x, ...)

  ## S3 method for class 'vector'
 addnames(x, prefix = "x", sep = "",
    ...)

  ## S3 method for class 'array'
 addnames(x,
    prefix = letters[1:length(dim(x))], sep = "", ...)

  ## S3 method for class 'matrix'
 addnames(x, prefix = c("row", "col"),
    ...)

Arguments

x
object whose names are generated.

prefix
prefix string to use. A vector can be used to specify a prefix for each dimension of x. Names are build as <prefix><sep><index>.

sep
separator used between the prefix and the numeric index.

...
extra arguments to allow extension and passed to the next method.

[Package pkgmaker version 0.22 Index]