Bibtex Utilities

Utility functions to work with BiBTeX files.

packageReferenceFile(PACKAGE = NULL, check = FALSE)

package_bibliography(PACKAGE = NULL, action = c("path", "copy", "load"))

Arguments

PACKAGE
package name. If NULL, then the name of the calling package is used.

check
logical that indicates if the result should be an empty string if the bibliography file (or package) does not exist.

action
single character string that specifies the action to be performed:

'path': return the path to the bibliography file. It returns an empty character string if the file does not exist.

'copy': copy the bibliography file to the current directory, overwriting any existing REFERENCES.bib file.

'load': load the bibliography file and return a list of utils::bibentry objects. It returns NULL if the file does not exist.

Functions

packageReferenceFile: returns the path to a package REFERENCES.bib file.

package_bibliography: returns the bibliography associated with a package. This can

Examples

packageReferenceFile('pkgmaker')

#> [1] "/tmp/RtmpJs7h8M/temp_libpath75ed85ee0ea45/pkgmaker/REFERENCES.bib"

packageReferenceFile('pkgmaker', check = TRUE)

#> [1] ""