cite_pkg {pkgmaker}	R Documentation

bibcite provides an inline package citation functionnality. Technically it adds a given Bibtex key to a cache that is used at the end of the document processing to generate a .bib file with all citation keys.

Description

bibcite provides an inline package citation functionnality. Technically it adds a given Bibtex key to a cache that is used at the end of the document processing to generate a .bib file with all citation keys.

Usage

  cite_pkg(key, cache = NA, ...)

Arguments

key
citation Bibtex key(s) as a character vector

cache
specifies what to do with the previsouly chached keys. If TRUE, then key is added to the cache. If NULL, then all previously cached keys are deleted, before . If a character string, then it specifies the path to a Bibtex file that is loaded to initialise the cache.

...
extra arguments passed to read.bib.

[Package pkgmaker version 0.22 Index]