citecmd {pkgmaker}	R Documentation

Citing Package References

Description

Create a citation command from package specific BibTex entries, suitable to be used in Rd files or Latex documents. The entries are looked in a file named REFERNCES.bib in the package's root directory (i.e. inst/ in development mode).

Usage

  citecmd(key, ..., REFERENCES = NULL)

Arguments

key
character vector of BibTex keys

...
extra arguments passed to format.bibentry.

REFERENCES
package or bibentry specification

Value

a character string containing the text formated BibTex entries

[Package pkgmaker version 0.22 Index]