fix_registry {pkgmaker}	R Documentation

Fix Registry Access Functions

Description

Fixes some of the member functions for registry objects, as defined in the registry package. The main fixed issue is due to the fact that key filtering does not return the correct entry, when an entry key is a prefix of another entry key, even when passed the exact key.

Usage

  fix_registry(regobj)

Arguments

regobj
registry object

[Package pkgmaker version 0.22 Index]