extractLocalFun {pkgmaker}	R Documentation

Extracting Local Function Definition

Description

extractLocalFun Extracts local function from wrapper functions of the following type, typically used in S4 methods: function(a, b, ...){ .local <- function(a, b, c, d, ...){} .local(a, b, ...) }

Works for methods that are created (setMethod) as a wrapper function to an internal function named .local.

Usage

  extractLocalFun(f)

  allFormals(f)

Arguments

f
definition of the wrapper function

Value

a function

a paired list like the one returned by formals.

[Package pkgmaker version 0.22 Index]