hasArg2 {pkgmaker}	R Documentation

Checking for Missing Arguments

Description

This function is identical to hasArg, except that it accepts the argument name as a character string. This avoids to have a check NOTE about invisible binding variable.

Usage

  hasArg2(name)

Arguments

name
the name of an argument as a character string.

Examples

f <- function(...){ hasArg2('abc') }
f(a=1)
f(abc=1)
f(b=1)

[Package pkgmaker version 0.22 Index]