install.dependencies {pkgmaker}	R Documentation

Installing All Package Dependencies

Description

Install all dependencies from a package source directory or package source file.

Usage

  install.dependencies(pkg = NULL, all = FALSE, ...,
    dryrun = FALSE)

Arguments

pkg
package path or source file

all
logical that indicates if 'Suggests' packages should be installed.

...
extra arguments passed to install.packages.

dryrun
logical that indicates if the packages should be effectively installed or only shown.

Examples

try( install.dependencies('Matrix', dryrun=TRUE) )
## Not run: 
install.dependencies("mypackage_1.0.tar.gz", dryrun=TRUE)

## End(Not run)

[Package pkgmaker version 0.22 Index]