list.libs {pkgmaker}	R Documentation

Library Files Utilities

Description

Lists binary library files in a directory

libname extracts library names from a path, removing the directory part of the path, as well as the platform specific library extension.

Usage

  list.libs(dir, ..., all.platforms = FALSE)

  libname(x)

Arguments

dir
directory

all.platforms
a logical that indicates whether to list library files for the current platform only (default) or all platforms (Unix, Windows, Mac).

...
extra arguments passed to list.files.

x
a filename

Value

a character vector

Examples

libname('mylib.so')
libname('/some/path/somewhere/mylib.dll')

[Package pkgmaker version 0.22 Index]