List Package Data Objects

Lists data objects that are shipped within package(s).

list.data(package = NULL)

Arguments

package
a single character string that specifies the name of a particular package where to look for data objects.

Value

a data.frame object with columns:

package: name of the package that holds the data object.

data: name of the key to use in utils::data or ldata to load the data object.

object: name of the (sub-)object that is contained in the data object.

See also

Examples

# list all data objects
head(list.data())

#>    package          data        object
#> 1 datasets AirPassengers AirPassengers
#> 2 datasets       BJsales       BJsales
#> 3 datasets       BJsales BJsales.lead 
#> 4 datasets           BOD           BOD
#> 5 datasets           CO2           CO2
#> 6 datasets   ChickWeight   ChickWeight

# list all data objects in package 'datasets'
subset(list.data("datasets"), data %in% "beavers")

#>     package    data   object
#> 46 datasets beavers beaver1 
#> 47 datasets beavers beaver2