onLoad {pkgmaker}	R Documentation

Default Load/Unload Functions

Description

Default Load/Unload Functions

Usage

  onLoad(libname = NULL, pkgname, chname = packageName())

  onUnload(libpath)

Arguments

libname
a character string giving the library directory where the package defining the namespace was found.

pkgname
a character string giving the name of the package.

libpath
a character string giving the complete path to the package.

chname
a character string naming a DLL (also known as a dynamic shared object or library) to load.

Examples

#----------
# onLoad
#----------
# in a package namespace:
.onLoad <- function(libname=NULL, pkgname){

	pkgmaker::onLoad(libname, pkgname)

}

#----------
# onUnload
#----------
# in a package namespace:
.onUnload <- function(libpath){

	pkgmaker::onUnload(libpath)

}

[Package pkgmaker version 0.22 Index]