Load Development Package

Load Development Package

load_project(
  pkg,
  reset = FALSE,
  ...,
  utests = TRUE,
  verbose = FALSE,
  addlib = TRUE,
  character.only = FALSE,
  try.library = FALSE
)

library_project(...)

Arguments

pkg
name of the package/project to load.

reset
logical that indicates if the package should be reloaded (passed to load_all.

...
other arguments passed to load_all.

utests
logical that indicates if an environment containing the unit test functions should be created. If TRUE this environment is accessible at pkgname::UnitTests$test.filename.r$function.name.

verbose
logical that indicates if log messages should be printed.

addlib
logical that indicates if the lib/ sub-directory, if it exists, should be prepended to the library path. This enables to control the version of the loaded dependencies.

character.only
logical that indicates if argument pkg should be evaluated or taken litteral.

try.library
logicatl that indicates if projects that could not be found should be looked up in the installed packages.

Functions

library_project: shortcut for load_project(..., try.library = TRUE), to load project code from installed library if not found as a development project. All its arguments are passed to load_project.