Logging Feature

lverbose returns/sets the current verbosity level.

lverbose(val, global = FALSE)

lsilent()

is.verbose()

lmessage(level, ..., appendLF = TRUE, sep = "", force = FALSE)

vmessage(...)

log_append(...)

Arguments

val
logical/numeric value that sets the verbosity level.

global
logical that indicates if the verbose level of all log handlers should be set to val.

level
verbosity level threshold (numeric value) above which the message should be printed out. This threshold is compared with the current verbosity level as returned by lverbose.

...
parts of a character message that are concatenated and passed to the current logger's write function.

appendLF
logical indicating if an endline character should be appended at the end of the message.

sep
separation character, used when concatenating all arguments in ....

force
logical that indicates if one should output messages or return a non null logger, even if the verbose mode is not high enough.

Value

the old verbose level

Functions

lsilent: tells if all verbose messages are silenced.

is.verbose: tells if verbosity is on, i.e. at level greater than 0.

lmessage: prints out a message (on sdtout) if the verbosity level is greater than a given value.

vmessage: prints out a log message (at level 1) using the current logger, typically on stdout. It is a shortcut for lmessage(1L, ...).

log_append: directly appends some message to the current log line.