lverbose {pkgmaker}	R Documentation

Internal verbosity option

Description

Internal verbosity option

Tells if all verbose messages should be

Tells if verbosity is on.

Prints out a message (on sdtout) if verbose mode is on.

Prints out a message (on sdtout) if the verbosity level is greater than a given value.

Usage

  lverbose(val)

  lsilent()

  is.verbose()

  vmessage(...)

  lmessage(..., level = 1L, appendLF = TRUE)

Arguments

val
logical that sets the verbosity level.

...
arguments passed to ... cat

appendLF
logical indicating if an endline character should be appended at the end of the message. Passed to cat.

level
verbosity level threshold (numeric value) above which the message should be printed out. This threshold is compared with the current verbosity level as returned by lverbose.

...
arguments passed to ... lmessage or cat

appendLF
logical indicating if an endline character should be appended at the end of the message. Passed to cat.

Value

the old verbose level

[Package pkgmaker version 0.22 Index]