makeUnitVignette {pkgmaker}	R Documentation

Make Vignette for Unit Tests

Description

Builds a vignette for unit tests in a package using the utest and a template vignette file.

Usage

  makeUnitVignette(pkg,
    file = paste(pkg, "-unitTests.pdf", sep = ""), ...,
    check = FALSE)

Arguments

pkg
Package name

file
Output file (.Rnw, .tex, or .pdf)

...
extra arguments passed to utest.

check
logical that indactes the cal was made from R CMD check, in which case the vignette is updated only if results of unit tests can be found in the unit test output directory, where they would have been generated by utest.

Value

Result of running unit test suite

[Package pkgmaker version 0.22 Index]