General Log Formating

Generate a formatted diagnostic message. This function is a shortcut for message(sprintf(...)).

messagef(fmt, ..., domain = NULL, appendLF = TRUE)

wnote(..., immediate. = TRUE)

Arguments

fmt
a character vector of format strings, each of up to 8192 bytes.

...
values to be passed into fmt. Only logical, integer, real and character vectors are supported, but some coercion will be done: see the ‘Details’ section. Up to 100.

domain
see gettext.

appendLF
logical: should messages given as a character string have a newline appended?

immediate.
logical, indicating if the call should be output immediately, even if getOption("warn") <= 0.

Functions

wnote: throws a simple note as an immediate warning. It is a shorcut for warning(..., immediate. = TRUE, call. = FALSE).

See also

Examples

messagef("Hello %s number %i", 'world', 4)

#> Hello world number 4