setBiocMirror {pkgmaker}	R Documentation

Setting Mirrors and Repositories

Description

setBiocMirror sets all Bioconductor repositories (software, data, annotation, etc.). so that they are directly available to install.packages. It differs from chooseBioCmirror in that it effectively enables the repositories.

getBiocMirror is a shortcut for getOption('BioC_mirror'), which returns the current Bioconductor mirror as used by biocLite.

getBiocRepos returns urls to all Bioconductor repositories on a given mirror.

setCRANMirror sets the preferred CRAN mirror.

CRAN simply contains the url of CRAN main mirror (http://cran.r-project.org), and aims at simplifying its use, e.g., in calls to install.packages.

Usage

  setBiocMirror(url = "http://www.bioconductor.org",
    version = NULL, unique = TRUE)

  getBiocMirror()

  getBiocRepos(url = "http://www.bioconductor.org",
    version = NULL)

  setCRANMirror(url = CRAN, unique = TRUE)

  CRAN

Arguments

url
or Bioconductor mirror url

version
version number

unique
logical that indicate if duplicated urls or names should be removed.

Format

chr "http://cran.r-project.org"

Examples

## Not run: 
install.packages('pkgmaker', repos=CRAN)

## End(Not run)

[Package pkgmaker version 0.22 Index]