getLoadingNamespace {pkgmaker}	R Documentation

Namespace Development Functions

Description

getLoadingNamespace returns information about the loading namespace. It is a wrapper to loadingNamespaceInfo, that does not throw an error.

Tests if a namespace is being loaded.

isNamespaceLoaded tests if a given namespace is loaded, without loading it, contrary to isNamespace.

isDevNamespace tests the – current – namespace is a devtools namespace.

Dynamically adds exported objects into the loading namespace.

ns_get gets an object from a given namespace.

Usage

  getLoadingNamespace(env = FALSE, info = FALSE,
    nodev = FALSE)

  isLoadingNamespace(ns, nodev = FALSE)

  isNamespaceLoaded(ns)

  isDevNamespace(ns)

  addNamespaceExport(x)

  ns_get(x, ns)

Arguments

env
logical that indicates that the namespace's environment (i.e. the namespace itself) should be returned.

info
logical that indicates that the complete information list should be returned

ns
the name of a namespace or a namespace whose loading state is tested. If missing isLoadingNamespace test if any namespace is being loaded.

nodev
logical that indicates if loading devtools namespace should be discarded.

x
character vector containing the names of R objects to export in the loading namespace.

Value

the name of the loading namespace if env and info are FALSE, an environment if env=TRUE, a list with elements pkgname and libname if info=TRUE.

[Package pkgmaker version 0.22 Index]