orderVersion {pkgmaker}	R Documentation

Ordering Version Numbers

Description

Orders a vector of version numbers, in natural order.

Usage

  orderVersion(x, decreasing = FALSE)

  sortVersion(x, ...)

Arguments

x
a character vector of version numbers

decreasing
a logical that indicates if the ordering should be decreasing

...
extra parameters passed to orderVersion

Examples

#----------
# orderVersion
#----------
v <- c('1.0', '1.03', '1.2')
order(v)
orderVersion(v)

#----------
# sortVersion
#----------
sort(v)
sortVersion(v)

[Package pkgmaker version 0.22 Index]