packageCLI {pkgmaker}	R Documentation

Package Specific Command Line Interface

Description

Package Specific Command Line Interface

Usage

  packageCLI(package, altfile = NULL, local = TRUE,
    ARGS = commandArgs(TRUE), ...)

Arguments

package
package name

altfile
alternative file that defines the main CLI entry point. That is a function named CLI, which takes the list of parsed command line arguments as its first argument.

local
logical that indicates if the main CLI function should be defined and evaluated in a local environment, or in the user's Global environment.

ARGS
list of parsed arguments passed to the main CLI function.

...
extra arugments passed to the package's CLI function.

[Package pkgmaker version 0.22 Index]