packageDependencies {pkgmaker}	R Documentation

List Package Dependencies

Description

List Package Dependencies

Usage

  packageDependencies(x, all = TRUE, as.list = FALSE,
    available = NULL)

Arguments

x
path to package source directory or file.

all
logical that indicates if all dependencies should be returned, or only the required ones.

as.list
logical that indicates if the result should be a list with one element per type of dependency.

available
a matrix of available packages (as returned by available.packages), from which the dependencies are retrieved. This means that there must be a row for the package x.

[Package pkgmaker version 0.22 Index]