packageReference {pkgmaker}	R Documentation

Package References

Description

Create a citation string from package specific BibTex entries, suitable to be used in Rd files. The entries are looked in a file named REFERNCES.bib in the package's root directory (i.e. inst/ in development mode).

Usage

  packageReference(key, short = FALSE)

Arguments

key
character vector of BibTex keys

short
logical that indicates if the reference should be shorten as First Author et al. if it has more than one author.

Value

a character string containing the text formated BibTex entries

[Package pkgmaker version 0.22 Index]