packageTestEnv {pkgmaker}	R Documentation

Returns the package internal environment where unit tests are stored.

Description

Returns the package internal environment where unit tests are stored.

Usage

  packageTestEnv(pkg)

Arguments

pkg
package name. If missing the caller's package is assumed.

[Package pkgmaker version 0.22 Index]