Loading Packages

require.quiet silently requires a package, and qrequire is an alias to require.quiet.

require.quiet(...)

qrequire(...)

qlibrary(...)

mrequire(msg, package, lib.loc = NULL, quietly = FALSE)

Arguments

...
extra arguments passed to library or require.

msg
error message to use, to which is appended the string ' requires package <pkg>' to build the error message.

package
name of the package to load.

lib.loc
a character vector describing the location of R library trees to search through, or NULL. The default value of NULL corresponds to all libraries currently known to .libPaths(). Non-existent library trees are silently ignored.

quietly
a logical. If TRUE, no message confirming package attaching is printed, and most often, no errors/warnings are printed if package attaching fails.

Functions

qlibrary: silently loads a package.

mrequire: tries loading a package with base require and stops with a -- custom -- error message if it fails to do so.

See also

Examples

mrequire('Running this example', 'stringr')

#> Loading required package: stringr

try( mrequire('Doing impossible things', 'notapackage') )

#> Loading required package: notapackage

#> Warning: there is no package called ‘notapackage’

#> Error in mrequire("Doing impossible things", "notapackage") : 
#>   Doing impossible things requires package 'notapackage'