quickinstall {pkgmaker}	R Documentation

Quick Installation of a Source Package

Description

Builds and install a minimal version of a package from its source directory.

Usage

  quickinstall(path, destdir = NULL, vignettes = FALSE,
    force = TRUE, ...,
    lib.loc = if (!is.null(destdir)) TRUE)

Arguments

path
path to the package source directory

destdir
installation directory. If NULL, the package is installed in the default installation library. If NA, the package is installed in a temporary directory, whose path is returned as a value.

vignettes
logical that indicates if the vignettes should be rebuilt and installed.

force
logical that indicates if the package should be installed even if a previous installation exists in the installation library.

...
extra arguments passed to R.CMD

lib.loc
library specification. If TRUE then the installation directory destdir is added to the default library paths. This can be usefull if dependencies are installed in this directory. If NULL, then the default library path is left unchanged.

Value

The path of the library where the package was installed.

[Package pkgmaker version 0.22 Index]