Renders rmarkdown Documents Using User Default Options

Renders rmarkdown Documents Using User Default Options

render_notes(
  input,
  output_format = NULL,
  output_options = NULL,
  ...,
  .config = NULL
)

Arguments

input
The input file to be rendered. This can be an R script (.R), an R Markdown document (.Rmd), or a plain markdown document.

output_format
The R Markdown output format to convert to. The option "all" will render all formats defined within the file. The option can be the name of a format (e.g. "html_document") and that will render the document to that single format. One can also use a vector of format names to render to multiple formats. Alternatively, you can pass an output format object (e.g. html_document()). If using NULL then the output format is the first one defined in the YAML frontmatter in the input file (this defaults to HTML if no format is specified there).

output_options
List of output options that can override the options specified in metadata (e.g. could be used to force self_contained or mathjax = "local"). Note that this is only valid when the output format is read from metadata (i.e. not a custom format object passed to output_format).

...
other arguments passed to render

.config
location of the default options (a YAML file). Default behaviour is to look for file '.rmarkdown.yaml' in the user's home directory, or, if missing, for a yaml section rmarkdown::render in the user's R profile.

See also