require.quiet {pkgmaker}	R Documentation

Silent Require

Description

Silently require a package.

Usage

  require.quiet(package, character.only = FALSE, ...)

Arguments

...
extra arguments passed to require.

package
the name of a package, given as a name or literal character string, or a character string, depending on whether character.only is FALSE (default) or TRUE).

character.only
a logical indicating whether package or help can be assumed to be character strings.

[Package pkgmaker version 0.22 Index]