setupPackageOptions {pkgmaker}	R Documentation

Package Specific Options

Description

The following functions to access/set the options from the set are assigned in envir:

<subset>Options

<subset>GetOption

Usage

  setupPackageOptions(..., NAME = NULL,
    ENVIR = topenv(parent.frame()),
    RESET = isLoadingNamespace())

Arguments

...
a single named list or named arguments that provide the default options and their values.

NAME
name of the set of options. This is used as a prefix for the name of the associated global option: package:<name>.

ENVIR
environment where the option wrapper functions will be defined. No function is defined if ENVIR=NULL

RESET
a logical that indicates whether the option set should overwrite one that already exists if necessary. The default is FALSE (i.e. no reset), except when loading a namespace, either from an installed package or a development package – with devtools. If FALSE, an error is thrown if trying to setup options with the same name.

[Package pkgmaker version 0.22 Index]