simpleRegistry {pkgmaker}	R Documentation

Simple Package Registry

Description

Simple Package Registry

Usage

  simpleRegistry(name, envir = topenv(parent.frame()),
    verbose = FALSE)

Arguments

name
name of the registry object, with which it will be assigned in envir.

envir
environment where to store the registry object. Defaults to the caller's top environment.

verbose
logical that toggle a verbose message when the object is first created.

[Package pkgmaker version 0.22 Index]