checkWarning {pkgmaker}	R Documentation

Extra Check Functions for RUnit

Description

checkWarning checks if a warning is generated by an expression, and optionally follows an expected regular expression pattern.

Usage

  checkWarning(expr, expected = TRUE, msg = NULL)

Arguments

expr
an R expression

expected
expected value as regular expression pattern. If a logical, then it specifies if a warning is expected or not.

For backward compatibility, a NULL value is equivalent to TRUE.

msg
informative message to add to the error in case of failure

Examples

# check warnings
checkWarning({ warning('ah ah'); 3})
checkWarning({ warning('ah oh ah'); 3}, 'oh')
try( checkWarning(3) )
try( checkWarning({ warning('ah ah'); 3}, 'warn you') )

[Package pkgmaker version 0.22 Index]