utest {pkgmaker}	R Documentation

Running Unit Tests

Description

Run unit tests in a variety of settings. This is still very experimental.

Usage

  utest(x, ...)

  ## S4 method for signature 'function'
utest(x, run = TRUE)

  ## S4 method for signature 'character'
utest(x,
    filter = "^runit.+\\.[rR]$", fun = "^test\\.", ...,
    testdir = "tests", framework = c("RUnit", "testthat"),
    quiet = Sys.getenv("RCMDCHECK") != "FALSE",
    lib.loc = NULL)

  ## S4 method for signature 'RUnitTestSuite'
utest(x, ..., quiet = FALSE,
    outdir = NULL)

Arguments

x
object to which a unit test is attached

...
extra arguments to allow extensions and are passed to the unit framework running funcitons.

run
a logical that indicates if the unit test should be run

filter
pattern to match files that contain the definition of the unit tests functions to run.

fun
patter to match the test functions to run.

testdir
directory where to look for the test files

framework
unit test framework

quiet
a logical that indicates if the tests should be run silently

lib.loc
path to a library where installed packages are searched for. Used is of the form x='package:*'.

outdir
output directory

Methods

utest

signature(x = "function"): Run the unit test assoicated to a function.

utest

signature(x = "character"): Run a package test suite

utest

signature(x = "RUnitTestSuite"): Runs a RUnit test suite

[Package pkgmaker version 0.22 Index]