utestFramework {pkgmaker}	R Documentation

Inferring Unit Test Framework

Description

Inferring Unit Test Framework

Usage

  utestFramework(x, eval = FALSE)

Arguments

x
an filename, a function or the body of a function

eval
a logical that indicates if the value of x should be used.

Value

the name of the framework as a character string or NULL if it could not be detected.

[Package pkgmaker version 0.22 Index]