utestPath {pkgmaker}	R Documentation

Unit Tests Result Directory

Description

Returns the path to the directory where the results of unit tests are stored. This path is used by utest to save unit test results, which are read by makeUnitVignette to update the unit test vignette when runnning R CMD check.

Usage

  utestPath(...)

Arguments

...
extra arguments passed to packagePath, e.g., package.

[Package pkgmaker version 0.22 Index]