writeUnitVignette {pkgmaker}	R Documentation

Writes Unit Tests Vignette

Description

Writes a vignette that contains the results from running unit test suites.

Usage

  writeUnitVignette(pkg, file, results = NULL,
    check = FALSE)

Arguments

pkg
Package name

file
Output Sweave (.Rnw) file

results
result file or output character vector

check
logical that indactes the cal was made from R CMD check, in which case the vignette is updated only if results of unit tests can be found in the unit test output directory, where they would have been generated by utest.

[Package pkgmaker version 0.22 Index]