write_PACKAGES_index {pkgmaker}	R Documentation

Generate CRAN-like Repository Index

Description

Generate CRAN-like Repository Index

Usage

  write_PACKAGES_index(path = ".", output = "index.html",
    pattern = NULL, title = "Packages", robots.file = TRUE)

Arguments

path
path to the repository's root directory

output
output filename – relative to the repository root path.

pattern
regular expression used to filter the names of the packages that will appear in the index.

title
title of the index page

robots.file
logical that indicates if a file robots.txt that hides the repository from search engine robots should be created.

[Package pkgmaker version 0.22 Index]