isCRANcheck {pkgmaker}	R Documentation

Package Check Utils

Description

isCRANcheck tries to identify if one is running CRAN-like checks.

isCRAN_timing tells if one is running CRAN check with flag '--timing'.

Currently, isCHECK checks both CRAN expected flags, the value of environment variable _R_CHECK_RUNNING_UTESTS_, and the value of option R_CHECK_RUNNING_EXAMPLES_. It will return TRUE if any of these environment variables is set to anything not equivalent to FALSE, or if the option is TRUE. For example, the function utest sets it to the name of the package being checked (_R_CHECK_RUNNING_UTESTS_=<pkgname>), but unit tests run as part of unit tests vignettes are run with _R_CHECK_RUNNING_UTESTS_=FALSE, so that all tests are run and reported when generating them.

Usage

  isCRANcheck(...)

  isCRAN_timing()

  isCHECK()

Arguments

...
each argument specifies a set of tests to do using an AND operator. The final result tests if any of the test set is true. Possible values are:

'timing'

Check if the environment variable _R_CHECK_TIMINGS_ is set, as with the flag '--timing' was set.

'cran'

Check if the environment variable _R_CHECK_CRAN_INCOMING_ is set, as with the flag '--as-cran' was set.

Details

Currently isCRANcheck returns TRUE if the check is run with either environment variable _R_CHECK_TIMINGS_ (as set by flag '--timings') or _R_CHECK_CRAN_INCOMINGS_ (as set by flag '--as-cran').

Important: the checks performed on CRAN check machines are – on purpose – not always run with such flags, so there is no guarantee this function effectively identifies such runs. CRAN recommends users rely on custom dedicated environment variables to enable specific tests or examples.

References

Adapted from the function CRAN in the fda package.

https://github.com/renozao/roxygen2

Examples

isCHECK()

[Package pkgmaker version 0.22 Index]