latex_preamble {pkgmaker}	R Documentation

LaTeX Utilities for Vignettes

Description

latex_preamble outputs/returns command definition LaTeX commands to be put in the preamble of vignettes.

latex_bibliography prints or return a LaTeX command that includes a package bibliography file if it exists.

Usage

  latex_preamble(PACKAGE, R = TRUE, CRAN = TRUE,
    Bioconductor = TRUE, GEO = TRUE, ArrayExpress = TRUE,
    biblatex = FALSE, only = FALSE, file = "")

  latex_bibliography(PACKAGE, file = "")

Arguments

R
logical that indicate if general R commands should be added (e.g. package names, inline R code format commands)

CRAN
logical that indicate if general CRAN commands should be added (e.g. CRAN package citations)

Bioconductor
logical that indicate if general Bioconductor commands should be added (e.g. Bioc package citations)

GEO
logical that indicate if general GEOmnibus commands should be added (e.g. urls to GEO datasets)

ArrayExpress
logical that indicate if general ArrayExpress commands should be added (e.g. urls to ArrayExpress datasets)

biblatex
logical that indicates if a \bibliography command should be added to include references from the package's REFERENCES.bib file.

only
a logical that indicates if the only the commands whose dedicated argument is not missing should be considered.

file
connection where to print. If NULL the result is returned silently.

PACKAGE
package name

Details

Argument PACKAGE is not required for latex_preamble, but must be correctly specified to ensure biblatex=TRUE generates the correct bibliography command.

Examples

latex_preamble()
latex_preamble(R=TRUE, only=TRUE)
latex_preamble(R=FALSE, CRAN=FALSE, GEO=FALSE)
latex_preamble(GEO=TRUE, only=TRUE)

[Package pkgmaker version 0.22 Index]