option_symlink {pkgmaker}	R Documentation

option_symlink creates a symbolic link to option x.

Description

option_symlink creates a symbolic link to option x.

is_option_symlink tests if x is a symbolic link option.

option_symlink_target returns the end target option of a symbolic link option x.

as.package_options creates an object such as the ones used to stores package specific options.

The method [[ is equivalent to options() or getOption(...): e.g. obj[[]] returns the list of options defined in obj, and obj[['abc']] returns the value of option 'abc'.

packageOptions provides access to package specific options from a given package that were defined with setupPackageOptions, and behaves as the base function options.

listPackageOptions returns the names of all option currently defined with setupPackageOptions.

Usage

  option_symlink(x)

  is_option_symlink(x, opts)

  option_symlink_target(x, opts)

  as.package_options(..., defaults = NULL)

  ## S3 method for class 'package_options'
 x[[...]]

  packageOptions(..., PACKAGE = packageName())

  listPackageOptions()

Arguments

opts
a list of options

x
a character string, a list or an object of class package_options.

defaults
NULL or a list of default options with their values.

...
arguments passed to getOption (only first one is used).

PACKAGE
a package name

Value

a character vector (possibly empty).

Examples

listPackageOptions()

[Package pkgmaker version 0.22 Index]