packageData {pkgmaker}	R Documentation

Loading Package Data

Description

Loads package data using data, but allows the user to avoid NOTEs for a ‘non visible binding variable’ to be thrown when checking a package. This is possible because this function returns the loaded data.

ldata loads a package data in the parent frame. It is a shortcut for packageData(list, ..., envir=parent.frame()).

Usage

  packageData(list, envir = .GlobalEnv, ...)

  ldata(list, ...)

Arguments

list
character vector containing the names of the data to load.

...
other arguments eventually passed to data.

envir
the environment where the data should be loaded.

Value

the loaded data.

Examples

#----------
# packageData
#----------
## Not run:  mydata <- packageData('mydata') 

#----------
# ldata
#----------
## Not run: 
# in a package' source => won't issue a NOTE
myfunction function(){
	mydata <- ldata('mydata')
}

## End(Not run)

[Package pkgmaker version 0.22 Index]