regfetch {pkgmaker}	R Documentation

Finds an entry in a registry.

Description

This function provides extra control on how entries are queried from a registry object.

pkgreg_fetch loads the requested package registry and uses regfetch to retrieve data from it.

pkgreg_remove removes an entry from a package registry.

Usage

  regfetch(regobj, ..., all = FALSE, error = TRUE,
    exact = FALSE, KEYS = NULL, verbose = FALSE,
    entry = FALSE, msg = NULL)

  pkgreg_fetch(regname, ..., msg = NULL,
    where = topenv(parent.frame()))

  pkgreg_remove(regname, ..., msg = NULL,
    where = topenv(parent.frame()), quiet = FALSE)

Arguments

regobj
a registry object

...
key value(s) to look up. If multiple indexes are used, then the primary key should come first.

all
logical to indicate if hidden keys (starting with a '.') should be returned and output in message.

error
a logical that indicates if an error should be thrown if the key has no match or multiple matches

exact
a logical that indicates if matching should be exact or partial. Note that if exact matches exist then they are returned, independently of the value of exact.

KEYS
alternative way of passing the key value(s). If not missing, then arguments in ... are discarded.

verbose
a logical that indicates if verbosity should be toggle on

entry
a logical that indicates if the

msg
a header to use in case of error.

quiet
a logical that indicates if the operation should be performed quietly, without throwing errors or warnings.

regname
Name of a sub-registry, used as its identifier.

where
package name or namespace that owns the registry.

[Package pkgmaker version 0.22 Index]